Embed a public API ABI jar in every distribution
Kotlin DSL script compilation only avoided generating the large
gradle-api.jar in :distributions-full, the sole distribution that
embedded the public API ABI jar. Every other distribution fell back
to generating it, defeating the benefit for those distributions.

Each distribution now produces its own ABI jar scoped to its runtime,
so the runtime fallback can be removed and script compilation resolves
the embedded jar in all distributions.

Split gradlebuild.public-api-jar so jar production is reusable by the
distributions, while publication stays specific to :public-api.
