+++ build-logic/packaging/src/main/kotlin/gradlebuild.public-api-jar.gradle.kts
@@ -15,182 +15,33 @@
  */
 
 import gradlebuild.basics.PublicApiVariants
-import gradlebuild.configureAsApiElements
 import gradlebuild.configureAsRuntimeJarClasspath
-import gradlebuild.packaging.support.ArtifactViewHelper.lenientProjectArtifactReselection
-import gradlebuild.packaging.support.FileLinesValueSource
-import gradlebuild.packaging.tasks.PathPrefixLister
-import org.gradle.kotlin.dsl.support.serviceOf
+import gradlebuild.packaging.support.includePublicApiAbiStubs
+import gradlebuild.packaging.support.publicApiAbiStubs
 
 plugins {
     id("gradlebuild.dependency-modules")
     id("gradlebuild.repositories")
     id("gradlebuild.reproducible-archives")
     id("gradlebuild.module-identity")
-    id("maven-publish")
-    id("signing")
 }
 
-description = "Generates a public API jar and corresponding component to publish it"
+description = "Produces the public API ABI jar (signature stubs) of the Gradle distribution it is applied to"
 
-// Defines configurations used to resolve external dependencies
-// that the public API depends on.
-// TODO: We should be able to derive these dependencies automatically.
-//       In fact, our public API should have no external dependencies.
-val externalApi = configurations.dependencyScope("externalApi") {
-    description = "External dependencies that the public Gradle API depends on"
-}
-val externalRuntimeOnly = configurations.dependencyScope("externalRuntimeOnly") {
-    dependencies.add(project.dependencies.create(project.dependencies.platform(project(":distributions-dependencies"))))
-}
-val externalRuntimeClasspath = configurations.resolvable("externalRuntimeClasspath") {
-    extendsFrom(externalApi.get())
-    extendsFrom(externalRuntimeOnly.get())
-    configureAsRuntimeJarClasspath(objects)
-}
-
-// Defines configurations used to resolve external dependencies that the legacy public API depends on.
-// The legacy variant inherits everything in the public API, minus the relocated impldeps qdox and javaparser-core.
-val legacyExternalApi = configurations.dependencyScope("legacyExternalApi") {
-    description = "External dependencies that the legacy public Gradle API depends on"
-    extendsFrom(externalApi.get())
-    exclude(module = "qdox")
-    exclude(module = "javaparser-core")
-}
-val legacyExternalRuntimeOnly = configurations.dependencyScope("legacyExternalRuntimeOnly") {
-    dependencies.add(project.dependencies.create(project.dependencies.platform(project(":distributions-dependencies"))))
-}
-val legacyExternalRuntimeClasspath = configurations.resolvable("legacyExternalRuntimeClasspath") {
-    extendsFrom(legacyExternalApi.get())
-    extendsFrom(legacyExternalRuntimeOnly.get())
-    configureAsRuntimeJarClasspath(objects)
-}
-
-// Defines configurations used to resolve the public Gradle API.
+// The source of API stubs to assemble into the ABI jar. Consumers populate it differently:
+// `:public-api` adds a dependency on the distribution to extract the full API surface from;
+// each distribution extends it from its own runtime buckets to extract only its own surface.
 val distribution = configurations.dependencyScope("distribution") {
     description = "Dependencies to extract the public Gradle API from"
 }
-
-// Resolvable configuration to get all projects having a runtime JAR
 val distributionClasspath = configurations.resolvable("distributionClasspath") {
     extendsFrom(distribution.get())
     configureAsRuntimeJarClasspath(objects)
 }
 
-fun Jar.commonPublicApiJarConfiguration() {
-    // We use the resolvable configuration, but leverage withVariantReselection to obtain the subset of api stubs artifacts
-    // Some projects simply don't have one, which excludes them
-    from(lenientProjectArtifactReselection<FileCollection>(
-        distributionClasspath,
-        {
-                attribute(Category.CATEGORY_ATTRIBUTE, objects.named("api-stubs"))
-            }
-    )) {
-        // TODO Use better filtering
-        include("**/*.class")
-        include("META-INF/*.kotlin_module")
-    }
-    // This is needed because of the duplicate package-info.class files, it is safe thanks to PackageInfoTest
-    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-}
-
-// Both jars carry the ABI of internals; the suffix marks the variant. The -legacy jar mirrors the
-// historical gradleApi() classpath and ships in lib/api/, pending an internals-free public API.
-val internalBaseName = gradleModule.identity.baseName.map { "$it${PublicApiVariants.INTERNAL_SUFFIX}" }
-val legacyBaseName = gradleModule.identity.baseName.map { "$it${PublicApiVariants.LEGACY_SUFFIX}" }
-
-val apiJarTask = tasks.register<Jar>("jarGradleApi") {
-    commonPublicApiJarConfiguration()
-    archiveBaseName = internalBaseName
-    destinationDirectory = layout.buildDirectory.dir("public-api/gradle-api")
-}
-
-val legacyApiJarTask = tasks.register<Jar>("jarGradleApiLegacy") {
-    commonPublicApiJarConfiguration()
-    archiveBaseName = legacyBaseName
+// Named after the runtime module so the module registry finds it as `gradle-public-api-legacy` in every distribution.
+tasks.register<Jar>("jarGradleApiLegacy") {
+    includePublicApiAbiStubs(publicApiAbiStubs(distributionClasspath))
+    archiveBaseName = PublicApiVariants.LEGACY_MODULE_NAME
     destinationDirectory = layout.buildDirectory.dir("public-api/gradle-api-legacy")
-}
-
-// The consumable configuration containing the public Gradle API artifact
-// and its external dependencies.
-val gradleApiElements = configurations.consumable("gradleApiElements") {
-    extendsFrom(externalApi.get())
-    outgoing.artifact(apiJarTask)
-    outgoing.capability(internalBaseName.map { "$group:$it:$version" })
-    configureAsApiElements(objects)
-}
-
-val legacyGradleApiElements = configurations.consumable("legacyGradleApiElements") {
-    extendsFrom(legacyExternalApi.get())
-    outgoing.artifact(legacyApiJarTask)
-    outgoing.capability(legacyBaseName.map { "$group:$it:$version" })
-    configureAsApiElements(objects)
-}
-
-val pathPrefixListTask = tasks.register<PathPrefixLister>("generateGradleApiPathPrefixList") {
-    apiJar.set(apiJarTask.flatMap { it.archiveFile} )
-    pathPrefixFile.set(layout.buildDirectory.file("public-api/path-prefix-list.txt"))
-}
-
-val sourceJarTask = tasks.register<Jar>("sourcesJar") {
-    val pathPrefixes = providers.of(FileLinesValueSource::class.java) {
-        parameters.inputFile.set(pathPrefixListTask.flatMap { it.pathPrefixFile })
-    }
-
-    val archiveOperations = project.serviceOf<ArchiveOperations>()
-    // We use the resolvable configuration, but leverage withVariantReselection to obtain the subset of sources
-    from(lenientProjectArtifactReselection(
-        distributionClasspath,
-        {
-                attribute(Category.CATEGORY_ATTRIBUTE, objects.named("documentation"))
-                attribute(DocsType.DOCS_TYPE_ATTRIBUTE, objects.named("sources"))
-            },
-        { fileCollection -> fileCollection.elements.map { set -> set.map { file -> archiveOperations.zipTree(file.asFile) } }}
-    )) {
-        include { element ->
-            val parent = element.relativePath.parent
-            if (parent == null) false
-            else pathPrefixes.get().contains(parent.pathString)
-        }
-    }
-
-    inputs.files(pathPrefixListTask.flatMap { it.pathPrefixFile }).withPropertyName("pathPrefixList").withPathSensitivity(PathSensitivity.NONE)
-    archiveClassifier.set("sources")
-    destinationDirectory = layout.buildDirectory.dir("public-api/gradle-api")
-    // This is needed because of the duplicate package-info.class files
-    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-}
-
-// TODO: SoftwareComponentFactoryProvider can be replaced with PublishingExtension#getSoftwareComponentFactory()
-open class SoftwareComponentFactoryProvider @Inject constructor(val factory: SoftwareComponentFactory)
-
-val softwareComponentFactory = project.objects.newInstance(SoftwareComponentFactoryProvider::class.java).factory
-val gradleApiComponent = softwareComponentFactory.adhoc("gradleApi")
-components.add(gradleApiComponent)
-
-// Published component containing the public Gradle API
-gradleApiComponent.addVariantsFromConfiguration(gradleApiElements) {
-    mapToMavenScope("compile")
-}
-
-val sourcesElements = configurations.consumable("sourcesElements") {
-    outgoing.artifact(sourceJarTask)
-    outgoing.capability(internalBaseName.map { "$group:$it:$version" })
-    attributes {
-        attribute(Category.CATEGORY_ATTRIBUTE, named(Category.DOCUMENTATION))
-        attribute(DocsType.DOCS_TYPE_ATTRIBUTE, named(DocsType.SOURCES))
-        attribute(Bundling.BUNDLING_ATTRIBUTE, named(Bundling.EXTERNAL))
-        attribute(Usage.USAGE_ATTRIBUTE, named(Usage.JAVA_RUNTIME))
-    }
-}
-
-gradleApiComponent.addVariantsFromConfiguration(sourcesElements) {
-    mapToOptional()
-}
-
-// Published component containing the legacy public Gradle API
-val legacyGradleApiComponent = softwareComponentFactory.adhoc("legacyGradleApi")
-components.add(legacyGradleApiComponent)
-legacyGradleApiComponent.addVariantsFromConfiguration(legacyGradleApiElements) {
-    mapToMavenScope("compile")
 }
