+++ platforms/jvm/distributions-jvm/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("gradlebuild.distribution.packaging")
+    id("gradlebuild.public-api-jar")
 }
 
 description = "The collector project for the 'jvm' portion of the Gradle distribution"
