+++ platforms/native/distributions-native/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("gradlebuild.distribution.packaging")
+    id("gradlebuild.public-api-jar")
 }
 
 description = "The collector project for the 'native' portion of the Gradle distribution"
